Extract loader/queries/structural/header into hooks (DocumentKnowledgeBase)
Second pass on #1446. Combined with the previous SidebarTabs/DocumentViewer/
useResizeHandle/useDocumentMarkdown extractions, DocumentKnowledgeBase.tsx
now sits at 1,057 lines — down from 2,654 (60% reduction). All public API
unchanged: same default export, same props, same rendered DOM, same
data-testids.

New extractions (sibling files in document_kb/):

- useDocumentLoader.ts: Owns every async fetch driving the view —
  GET_DOCUMENT_KNOWLEDGE_AND_ANNOTATIONS, GET_DOCUMENT_WITH_STRUCTURE,
  GET_DOCUMENT_ANNOTATIONS_ONLY, GET_CONVERSATIONS — plus the
  processAnnotationsData / processAnnotationsOnlyData / loadDocxDocument /
  PDF-PAWLs page-builder helpers and the analysis/extract refetch effects.
  Both query paths now share a single loadDocumentBody dispatcher; the
  corpus path threads through the cached-URL helpers, the no-corpus path
  goes direct, but the rest of the PDF/TXT/DOCX branches are identical.
  Returns just what the parent needs: combinedData, viewState, loading,
  queryError, refetch, threadCount.
- useStructuralAnnotations.ts: Lazy-loads the structural-annotation set
  (toggle path) and targeted deep-link fetches. Pure side-effect hook;
  manages atoms internally. Resets atoms on documentId change.
- useContainerWidth.ts: Container width tracking + ResizeObserver +
  scrollContainerRefAtom wiring.
- HeaderBar.tsx: Modal title row + metadata + Add-to-Corpus / back
  buttons. Stateless presentation.
- helpers.ts: Pure relationToGroup factory shared between the loader and
  structural-annotation hooks (was duplicated inline at three call sites).

State doesn't need to be lifted to a Jotai atom for this slice — every
extracted hook either already uses atoms internally or takes the props
the parent already had on hand.

Verified:
- tsc --noEmit clean
- prettier formatted
- Component tests pass: Index/chat/feed/extract/analysis tab visibility,
  PDF/TXT/DOCX viewer rendering, analysis selection opens sidebar with
  Analysis tab, extract selection opens sidebar with Extract tab, context
  bar shows correct stats, document renders with and without corpus,
  Add to Corpus modal.
